requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
fyers-apiv3>=2.1.0
python-dotenv>=1.0.0
plotly>=5.0.0
ta>=0.10.0
cryptography>=41.0.0
supabase>=2.0.0
openpyxl>=3.11.0
pytz>=2023.3
bcrypt>=4.0.0
